requirements.txt
streamlit==1.10.0
google-generativeai
dotenv